% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevAnalysis.R
\name{classificazione_reintegri}
\alias{classificazione_reintegri}
\title{classificazione_reintegri}
\usage{
classificazione_reintegri()
}
\value{
La funzione produce come output primario un file Excel che categorizza i vari bonifici di reintegro effettuati dal dealer nell'arco di tempo specificato dal file importato originariamente
}
\description{
La funzione produce come output un file Excel che riepiloga e categorizza i vari bonifici di reintegro effettuati dal dealer nell'arco di tempo specificato dal file importato originariamente
}
\examples{
classificazione_reintegri()
}
\author{
Anthony Tricarico
}
