% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevAnalysis.R
\name{clean}
\alias{clean}
\title{clean}
\usage{
clean()
}
\value{
la funzione produce come output un oggetto data.frame di R da assegnare alla variabile movimenti_cleaned
}
\description{
La funzione pulisce il file excel precedentemente importato permettendo anche di gestire due "accordati" diversi in base alla data di operatività della linea. La funzione va eseguita dopo aver eseguito import().
}
\examples{
movimenti_cleaned <- clean()
}
\author{
Anthony Tricarico
}
