% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevAnalysis.R
\name{import}
\alias{import}
\title{Import}
\usage{
import()
}
\value{
la funzione produce come output un oggetto data.frame di R da assegnare alla variabile import
}
\description{
La funzione importa il file excel per l'analisi eseguendo automaticamente tutti i passaggi per preparare i dati alla corretta gestione nei passaggi successivi
}
\details{
è importante fornire il nome del file da importare in maniera corretta esattamente come appare nella cartella dove è stato salvato
}
\examples{
import <- import()
}
\author{
Anthony Tricarico
}
