% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevAnalysis.R
\name{utilizzo_daily}
\alias{utilizzo_daily}
\title{utilizzo_daily}
\usage{
utilizzo_daily()
}
\value{
La funzione produce come output primario un file Excel che permette di vedere il saldo giornaliero del dealer in oggetto.
}
\description{
La funzione genera il saldo giornaliero dell'utilizzo della linea Revolving andando a stimare un utilizzo giornaliero progressivo della linea.
La funzione concede, inoltre, la possibilità di generare un grafico dell'utilizzo della linea su base giornaliera nell'intervallo di tempo specificato.
}
\examples{
utilizzo_daily()
}
\author{
Anthony Tricarico
}
