% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevAnalysis.R
\name{valutazione_reintegri_saldo_attivo}
\alias{valutazione_reintegri_saldo_attivo}
\title{valutazione_reintegri_saldo_attivo}
\usage{
valutazione_reintegri_saldo_attivo()
}
\value{
La funzione produce come output primario un file Excel che permette di vedere quali bonifici di reintegro sono stati effettuati in presenza di saldo attivo
}
\description{
La funzione produce come output un file Excel che permette di isolare i bonifici di reintegro che sono stati effettuati con saldo attivo
}
\examples{
valutazione_reintegri_saldo_attivo()
}
\author{
Anthony Tricarico
}
